pyproject.toml
[tool.poetry]
name = "innovation"
version = "0.1.0"
description = ""
authors = ["Martí Umbert <marti.umbert@bsc.es>"]
readme = "README.md"
packages = [
    { include = "innovation" }
]

include = [
    "innovation/gendata/tasks_examples/*",
    "innovation/gendata/models_examples/*"
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyyaml = {version = ">=6.0.2,<7.0.0", optional=false}
torch = {version = ">=2.6.0,<3.0.0", optional=false}

openai = {version = "1.74.0", optional=true}
pandas = {version = "2.2.3", optional=true}
faster-whisper = {version = ">=1.1.1,<2.0.0", optional=true}
tqdm =  {version = ">=4.67.1,<5.0.0", optional=true}

[tool.poetry.extras]
speech = ["tqdm", "faster-whisper"]
gendata = ["pandas", "openai"]

[build-system]
requires = [
  "poetry-core>=2.0.0,<3.0.0",
  "poetry-plugin-export>=1.8.0"
]
build-backend = "poetry.core.masonry.api"
